+++ ChangeLog.md
@@ -1,5 +1,11 @@
 # Changelog for PiG
 
+### 0.15.1
+* implement code completion
+
+### 0.15.0
+* change order of monads in Interp to allow for auto completion
+
 ### 0.14.2
 * change `--file` to `--load` to make it more consistent with language